new small features
made progress bar as thick as the card and somewhat transparent, looks nicer
made the download progress not cut off and be seamless when transitioning from intermediate to solid colors
renamed history tab to downloads
added videopath to downloaded items and if they are not in the filesystem, they are greyed out
made updating app as a toggle in the main settings page. moved update ytdl there aswell
fixed not getting all playlist items from generic ytdl request
brought back the default api key. If the key is throttled, it will try to use the user's custom youtube api key. If that key is throttled too, it will use the generic request as a last resort