stuff
- added saved downloads in the settings backup
- made icon bigger in downloading notification
- fixed notification cancel and pausing not working
- fixed app not working in preferred home screen for downloads and more
- removed 000001 if the user has made only one cut
- removed log id from downloads if the user deleted the log
- if the user had multiple preferred audio formats the app would choose the last one and not the first
- added ability to combine preferred audio formats by writing like 140+251,250. If the download item finds both first audio formats it will merge them together, otherwise get one of them that is available
- fixed app file transfer not working. (had to revert to hidden cache folder as downloads folder wasnt good for some phones)
- fixed app saving as mkv even though mp4 is set as container
- added shortcuts button to the commands tab so you can drop shortcuts to the current command.
- added syntax highlighting in commands tab textbox
- fixed tiktok videos not saving properly
- added ability to combine all possible combinations of preferred video formats, audio formats, codex and container.

The app will consider all preferred video ids. For each of them it will consider preferred vcodec and extension. At the same time for each of those elements it will consider all preferred audio commands and/or audio merge.
If the format is a generic quality like 1080p or 720p it will also consider that in the long format query. Also the app will add all combos of video + audio only. And all combos of all video + best audio. All video ids alone and best as final fallback.
If the user chooses a proper format, the app will not be this descriptive. Preferred video id and audio id combos will only apply on a quick download or best quality generic download